requirements.txt
ffmpeg
ffmpeg-python
yt_dlp
firebase_admin
firebase-admin
python-dotenv
Flask